Appears to be fully running
